CatailystPythonTools/requirements.txt
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
idna==2.10
joblib==1.0.1
lxml==4.6.3
numpy==1.21.0
pandas==1.2.5
pandas-datareader==0.9.0
python-dateutil==2.8.1
pytz==2021.1
regex==2021.4.4
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
tqdm==4.61.1
urllib3==1.26.6
yahoofinancials==1.6
requests~=2.25.1
selenium~=3.141.0
beautifulsoup4~=4.9.3
newspaper3k==0.2.8